refactor: インラインテストを各 module/tests.rs に分離
rurico#201 のパターンを横展開。15 production module の
`#[cfg(test)] mod tests { ... }` を逐語移動して `module/tests.rs` に分離し、
本体は `#[cfg(test)] mod tests;` のみ残す。テスト名・assertion は不変。

対象 (16 候補中、main.rs は bin root のため除外):
config / error / fs_optional / injection_check / io / query_log / recall /
resolver / text / verify / tools/format / recall/corpus /
indexer/embed / indexer/walker / indexer/source_kind

唯一の適応: injection_check/tests.rs の `include_str!` 2 箇所を `../` 前置
(`tools/format.rs` -> `../tools/format.rs`、`brief.rs` -> `../brief.rs`)。tests.rs が
1 階層深くなり file-relative パスがずれるため。対象ファイル・assertion は不変。

検証: nextest list の test-name 集合が移動前後で完全一致 (713、追加/削除/改名ゼロ)、
nextest run 713 passed、clippy clean、fmt clean。

Closes #247
